Quan nd

+++ src/main/java/com/example/c4zone/model/product/Product.java
@@ -28,7 +28,7 @@ public class Product {
     private String selfieProduct;
     private String batteryProduct;
     private String weightProduct;
-    private Double quantityProduct;
+    private Long quantityProduct;
     private Double priceProduct;
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL)
     private List<Image> images;
